@@ -40,86 +37,53 @@
     }
   },
   "dependencies": {
-    "electron-is-dev": "^0.3.0",
-    "fuse.js": "^3.2.0",
-    "localforage": "^1.6.0",
-    "react": "^16.2.0",
-    "react-beautiful-dnd": "^6.0.0",
-    "react-dom": "^16.2.0",
+    "@types/lodash": "^4.14.112",
+    "fuse.js": "^3.2.1",
+    "humps": "^2.0.1",
+    "localforage": "^1.7.1",
+    "lodash": "^4.17.10",
+    "normalizr": "^3.2.4",
+    "react": "^16.3.2",
+    "react-beautiful-dnd": "^8.0.5",
+    "react-dom": "^16.3.2",
     "react-redux": "^5.0.7",
-    "react-relay": "^1.5.0",
     "react-router": "^4.2.0",
-    "react-router-config": "^1.0.0-beta.4",
-    "react-router-dom": "^4.2.2",
     "react-select": "^1.2.1",
-    "redux": "^3.7.2",
-    "redux-api": "^0.11.2",
-    "redux-form": "^7.3.0",
     "redux-persist": "^5.9.1",
-    "redux-thunk": "^2.2.0",
-    "relay-runtime": "^1.5.0"
+    "redux-saga": "^0.16.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0-beta.42",
-    "@babel/preset-env": "^7.0.0-beta.42",
-    "@babel/preset-react": "^7.0.0-beta.42",
-    "@babel/preset-typescript": "^7.0.0-beta.42",
-    "@types/react": "^16.0.40",
-    "@types/react-beautiful-dnd": "^4.0.0",
-    "@types/react-dom": "^16.0.4",
-    "@types/react-redux": "^5.0.15",
-    "@types/react-relay": "^1.3.4",
-    "@types/react-router-config": "^1.0.6",
-    "@types/react-router-dom": "^4.2.5",
-    "@types/react-select": "^1.2.5",
-    "@types/redux-form": "^7.2.2",
-    "@types/redux-persist": "^4.3.1",
-    "@types/redux-thunk": "^2.1.0",
-    "awesome-typescript-loader": "^5.0.0-1",
-    "babel-eslint": "^8.2.2",
-    "babel-loader": "^8.0.0-beta",
-    "babel-plugin-relay": "^1.5.0",
+    "@types/humps": "^1.1.2",
+    "@types/normalizr": "^2.0.18",
+    "@types/react": "^16.4.6",
+    "@types/react-beautiful-dnd": "^7.1.0",
+    "@types/react-dom": "^16.0.5",
+    "@types/react-redux": "^6.0.4",
+    "@types/react-relay": "^1.3.5",
+    "@types/react-router-config": "^1.0.8",
+    "@types/react-router-dom": "^4.3.0",
+    "@types/react-select": "^1.2.11",
+    "awesome-typescript-loader": "^5.0.0",
     "concurrently": "^3.5.1",
-    "copy-webpack-plugin": "^4.5.1",
-    "css-loader": "^0.28.11",
-    "dotenv-webpack": "^1.5.5",
-    "electron": "^1.8.4",
-    "electron-builder": "^20.6.2",
-    "electron-devtools-installer": "^2.2.3",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.19.1",
-    "eslint-plugin-react": "^7.7.0",
+    "copy-webpack-plugin": "^4.5.2",
+    "css-loader": "^1.0.0",
+    "electron": "^2.0.5",
+    "electron-builder": "^20.24.4",
+    "electron-debug": "^2.0.0",
+    "electron-devtools-installer": "^2.2.4",
+    "eslint": "^5.2.0",
+    "eslint-loader": "^2.1.0",
+    "eslint-plugin-react": "^7.10.0",
     "file-loader": "^1.1.11",
-    "graphql": "^0.13.2",
-    "graphql-cli": "^2.15.8",
-    "html-webpack-plugin": "^3.0.7",
-    "lint-staged": "^7.0.0",
-    "node-sass": "^4.7.2",
-    "relay-compiler": "^1.5.0",
-    "sass-loader": "^6.0.7",
-    "style-loader": "^0.20.3",
-    "typescript": "^2.7.2",
-    "typescript-eslint-parser": "^14.0.0",
-    "webpack": "^4.2.0",
-    "webpack-cli": "^2.0.12",
-    "webpack-serve": "^0.2.0"
-  },
-  "jest": {
-    "setupFiles": [
-      "./config/jest.setup.js"
-    ],
-    "moduleDirectories": [
-      "./node_modules"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx"
-    ],
-    "transform": {
-      "^.+\\.(js|jsx|ts|tsx)$": "typescript-babel-jest"
-    }
+    "html-webpack-plugin": "^3.2.0",
+    "lint-staged": "^7.1.0",
+    "node-sass": "^4.9.2",
+    "sass-loader": "^7.0.1",
+    "source-map-loader": "^0.2.3",
+    "style-loader": "^0.21.0",
+    "typescript": "^2.8.3",
+    "webpack": "^4.16.1",
+    "webpack-cli": "^3.1.0",
+    "webpack-serve": "^2.0.2"
   }
 }
