@@ -2,19 +2,20 @@ import * as React from 'react'
 import ExternalLinkIcon from 'views/ui/icons/ExternalLinkIcon'
 
 type ExternalLinkProps = {
-  showIcon: boolean;
+  children?: React.ReactNode;
+  className?: string;
+  showIcon?: boolean;
   url: string;
-  children: React.ReactNode;
-  className: string;
 }
 
 const handleClick = (e: any) => {
   e.preventDefault()
+  const url = e.currentTarget.getAttribute('href')
 
-  if (window.process) {
-    const shell = window.require('electron').shell
-    shell.openExternal(e.currentTarget.getAttribute('href'))
-  }
+  if (window.shell)
+    window.shell.openExternal(url)
+  // else
+  //   window.location = url
 }
 
 const ExternalLink: React.SFC<ExternalLinkProps> = (props) => {
